+++ src/component/index.ts
@@ -1,9 +1,5 @@
-export {
-  Data,
-  defineComponent,
-  SetupFunction,
-  SetupContext,
-  ComponentInstance,
-  ComponentRenderProxy,
-} from './component'
+export { defineComponent } from './defineComponent'
+export { SetupFunction, SetupContext } from './componentOptions'
+export { ComponentInstance, ComponentRenderProxy } from './componentProxy'
+export { Data } from './common'
 export { PropType, PropOptions } from './componentProps'